+++ pyraxshell/plugins/libservices.py
@@ -15,4 +15,6 @@
 # You should have received a copy of the GNU General Public License
 # along with pyraxshell. If not, see <http://www.gnu.org/licenses/>.
 
-VERSION='0.2.1'
+
+class LibServices(object):
+    pass