+++ pyraxshell/plugins/plugin_servers.py
@@ -31,18 +31,18 @@ def injectme(c):
     logging.debug('%s injected' % __file__)
 
 def do_servers(*args):
-    Cmd_Servers().cmdloop()
+    Cmd_servers().cmdloop()
 
 
-class Cmd_Servers(Plugin, cmd.Cmd):
+class Cmd_servers(Plugin, cmd.Cmd):
     '''
     pyrax shell POC - Manage servers module
     '''
     
-    prompt = "H servers>"    # default prompt
+    prompt = "RS servers>"    # default prompt
 
     def __init__(self):
-        cmd.Cmd.__init__(self)
+        Plugin.__init__(self)
         self.libplugin = LibServers()
 
     # ########################################
@@ -152,6 +152,15 @@ class Cmd_Servers(Plugin, cmd.Cmd):
                            for f in params
                             if f.startswith(text)
                             ]
+#TODO --
+        last_token = line.split()[-1]
+        logging.debug("last_token: %s" % last_token)
+        if last_token == 'flavor_id:':
+            logging.debug("auto-complete flavour")
+        elif last_token == 'image_id:':
+            logging.debug("auto-complete image_id")
+        elif last_token == 'name:':
+            logging.debug("auto-complete name")
         return completions
     
     def do_delete(self, line):
@@ -359,7 +368,7 @@ class Cmd_Servers(Plugin, cmd.Cmd):
         except:
             logging.error(traceback.format_exc())
     
-    def complete_take_snapshot(self, text, line, begidx, endidx):
+    def complete_take_snapshots(self, text, line, begidx, endidx):
         params = ['id:', 'snapshot_name:']
         if not text:
             completions = params[:]