+++ RELEASENOTES.md
@@ -1,6 +1,21 @@
 Release notes for pyraxshell
 =======
 
+## 2013.09.21 - Version 0.2.1
+  * Fixed issue #7 - Remove "unload_plugins"
+  * Fixed 'import' in 'pyraxshell::load_plugins()'
+  * Fixed issue #6
+  * Fixed issue #10
+
+## 2013.09.16 - Version 0.2.0
+  * Added features to plugin_loadbalancers
+  * Added features to plugin_services
+  * Added features to plugin_servers
+
+## 2013.09.15 - Version 0.1.9
+  * Minor fixes
+  * Fixed plugin_dns methods
+  
 ## 2013.09.14 - Version 0.1.8
   * Fixed issue #1
   * Added base Plugin class